Split C codegen tests
