+++ docs/PHASE_PLAN.md
@@ -189,6 +189,8 @@ checked-in docs, tests, and commits only.
 - Typechecker call signature and generic validation helpers now live in a
   dedicated expression helper module, keeping call-expression dispatch separate
   from argument, annotation, and return-flow validation.
+- C backend unit tests now live beside the backend in a dedicated test module,
+  keeping the production C backend entry point and identifier helpers compact.
 - Resolver validation replay now collects expected declaration symbols,
   expected local symbols, import-validation state, and behavior-association
   replay tasks in one declaration pass before checking resolver-owned extras,
